requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiokafka>=0.10.0
redis[hiredis]>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
structlog>=23.0.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
python-multipart>=0.0.6
websockets>=12.0
prometheus-client>=0.19.0
slowapi>=0.1.9
pytz>=2023.3
scipy>=1.11.0
alembic>=1.12.0
tenacity>=8.2.0
psutil>=5.9.0
aiodocker>=0.21.0
aiosmtplib>=2.0.0

# Firebase & Authentication
firebase-admin>=6.2.0
PyJWT>=2.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Performance optimizations
cachetools>=5.3.0
msgpack>=1.0.0
brotli>=1.1.0
lz4>=3.1.0

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0
grpcio-health-checking>=1.60.0
protobuf>=4.25.0
docker>=7.0.0
python-on-whales>=0.68.0
apscheduler>=3.10.0
fastavro>=1.9.0

# Observability & Distributed Tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-asyncpg==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
python-snappy>=0.6.0
